Fix Expand All/Minimize All buttons functionality and use hooks instead of HOC for context.
